refactor(board): 게시글 상세 화면 features 모듈 분할 (Phase 4)
- src/features/board/postDetail/(app/ 밖, Expo Router 라우트 오인 방지): Skeleton/CommentSectionHeader/InlineComposerRow/PostHeader 컴포넌트 + 순수 헬퍼 + useBoardPostDetailScreen 훅
- 훅 deps·ref·로직 그대로 이동, 렌더 트리·Provider 위치 불변
- 진입점 [postId].tsx(1180→270)는 훅 호출+JSX 조립+라우트 가드만, default export 불변
- renderItem 안정성 테스트 포함 board 전 테스트 통과 (tsc 0, test 4286 pass)
